+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 4.1 (unreleased)
 ----------------
 
+- Fix HTML quoting in test forms
+  (`#44 <https://github.com/zopefoundation/Products.ZSQLMethods/issues/44>`_)
+
 
 4.0 (2023-02-02)
 ----------------